# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Expense Tracker API
  version: 1.0.0
  description: |
    REST API for storing purchase transactions in USD and retrieving them
    converted into any currency reported by the US Treasury Reporting Rates
    of Exchange.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: transactions
    description: Purchase transaction lifecycle

paths:
  /api/v1/transactions:
    post:
      tags: [transactions]
      operationId: createTransaction
      summary: Store a purchase transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
              description: URI of the newly created transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [transactions]
      operationId: listTransactions
      summary: List purchase transactions (paged)
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Page of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionPage'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/transactions/{id}:
    get:
      tags: [transactions]
      operationId: getTransaction
      summary: Retrieve a transaction, optionally converted to a target currency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: currency
          in: query
          required: false
          description: |
            ISO 4217 currency code. When supplied, the response includes
            `exchangeRate`, `convertedAmount` and `rateDate` using the most
            recent Treasury rate dated on or before `transactionDate` within
            the prior 6 months. If no such rate exists, 422 is returned.
          schema:
            type: string
            pattern: '^[A-Z]{3}$'
            example: BRL
      responses:
        '200':
          description: Transaction (optionally enriched with conversion)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Transaction'
                  - $ref: '#/components/schemas/ConvertedTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /api/v1/transactions/{id}/cache:
    delete:
      tags: [transactions]
      operationId: evictTransactionCache
      summary: Evict cached FX rates for a transaction
      description: |
        Removes all cached FX rate entries for the transaction's date quarter,
        forcing the next conversion request to fetch a fresh rate from the
        Treasury API. Returns 204 whether or not any entries were cached.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Cache entries evicted (no-op if none were cached)
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    PageNumber:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    MonetaryAmount:
      type: string
      description: |
        Decimal amount with up to 2 fractional digits, serialised as a string
        to preserve precision. Always non-negative on input.
      pattern: '^(0|[1-9][0-9]*)(\.[0-9]{1,2})?$'
      example: '125.49'

    ExchangeRate:
      type: string
      description: |
        Treasury exchange rate (target units per USD) with up to 6 fractional
        digits, serialised as a string.
      pattern: '^[0-9]+(\.[0-9]{1,6})?$'
      example: '5.123400'

    IsoDate:
      type: string
      format: date
      example: '2026-05-15'

    CurrencyCode:
      type: string
      description: ISO 4217 alphabetic currency code.
      pattern: '^[A-Z]{3}$'
      example: BRL

    CreateTransactionRequest:
      type: object
      required: [description, transactionDate, purchaseAmountUsd]
      additionalProperties: false
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 50
          example: Office supplies
        transactionDate:
          $ref: '#/components/schemas/IsoDate'
        purchaseAmountUsd:
          $ref: '#/components/schemas/MonetaryAmount'

    Transaction:
      type: object
      required: [id, description, transactionDate, purchaseAmountUsd]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 50
        transactionDate:
          $ref: '#/components/schemas/IsoDate'
        purchaseAmountUsd:
          $ref: '#/components/schemas/MonetaryAmount'

    ConvertedTransaction:
      type: object
      required:
        [
          id,
          description,
          transactionDate,
          purchaseAmountUsd,
          currency,
          exchangeRate,
          convertedAmount,
          rateDate,
        ]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 50
        transactionDate:
          $ref: '#/components/schemas/IsoDate'
        purchaseAmountUsd:
          $ref: '#/components/schemas/MonetaryAmount'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        exchangeRate:
          $ref: '#/components/schemas/ExchangeRate'
        convertedAmount:
          $ref: '#/components/schemas/MonetaryAmount'
        rateDate:
          $ref: '#/components/schemas/IsoDate'

    TransactionPage:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          minimum: 0
          format: int64
        totalPages:
          type: integer
          minimum: 0

    Problem:
      type: object
      description: RFC 7807 problem+json error response.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errors:
          type: array
          description: Field-level validation errors when applicable.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string

  responses:
    BadRequest:
      description: Validation failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: Request was syntactically valid but cannot be processed (e.g. no FX rate within 6 months).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'